revert changes with serializer
